validation

+++ src/main/java/tn/esprit/pidevcrashcode/Repositories/UserRepository.java
@@ -1,15 +1,26 @@
 package tn.esprit.pidevcrashcode.Repositories;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import tn.esprit.pidevcrashcode.Entities.Activity;
+import tn.esprit.pidevcrashcode.Entities.Location;
+import tn.esprit.pidevcrashcode.Entities.TypeActivity;
 import tn.esprit.pidevcrashcode.Entities.User;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     Optional<User> findByUsername(String username);
 
     Optional<User> findByEmail(String email);
+
+
+
+
+
+
 }
